Makes Manual Bail Respawn Compatible with Respawn Near Bail
Adds compatibility between these two mods. Previously you'd do the near bail respawn, and then press A and respawn at the marker. This makes it so if Respawn Near Bail is active it uses its respawn method, instead of the default, and cancels the coroutine that manually does it after 2.5 seconds.